requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
attrs==24.3.0
autopep8==2.3.2
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
cssselect2==0.7.0
Django==5.1.5
django-cors-headers==4.6.0
django-rest-swagger==2.2.0
django-silk==5.3.2
django-simple-history==3.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.8
gprof2dot==2024.6.6
html5lib==1.1
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.0
MarkupSafe==3.0.2
openapi-codec==1.3.2
oscrypto==1.3.0
packaging==24.2
pillow==11.1.0
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pycparser==2.22
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyJWT==2.10.1
pypdf==5.1.0
python-bidi==0.6.3
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.1
reportlab==4.2.5
requests==2.32.3
rpds-py==0.22.3
simplejson==3.19.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
typing_extensions==4.12.2
tzlocal==5.2
uritemplate==4.1.1
uritools==4.0.3
urllib3==2.3.0
webencodings==0.5.1
xhtml2pdf==0.2.16
